backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain_anthropic
tavily-python
langchain_community